Share one formatFileSize between the backup surfaces
