+++ client/packages/cli/src/commands/backup/list.ts
@@ -1,26 +1,11 @@
 import chalk from 'chalk';
 import { Effect } from 'effect';
-import type { AppBackup } from '@instantdb/platform';
+import { formatFileSize, type AppBackup } from '@instantdb/platform';
 import type { backupListDef, OptsFromCommand } from '../../index.ts';
 import { useBackupsManager } from '../../lib/backups.ts';
 
 export const formatBackupDate = (date: Date) =>
   `${date.toISOString().replace('T', ' ').slice(0, 16)} UTC`;
-
-export function formatBytes(n: number): string {
-  // Decimal (1000-based) units with SI labels, to match how macOS/Finder
-  // reports file sizes.
-  if (n < 1000) return `${n} B`;
-  const units = ['KB', 'MB', 'GB', 'TB'];
-  let i = -1;
-  let v = n;
-  do {
-    v /= 1000;
-    i++;
-  } while (v >= 1000 && i < units.length - 1);
-  const digits = v < 10 ? 2 : v < 100 ? 1 : 0;
-  return `${v.toFixed(digits)} ${units[i]}`;
-}
 
 export const renderBackup = (backup: AppBackup) =>
   Effect.gen(function* () {
@@ -30,11 +15,11 @@ export const renderBackup = (backup: AppBackup) =>
       yield* Effect.log(`  Description: ${backup.description}`);
     }
     if (backup.dbSize != null) {
-      yield* Effect.log(`  Database size: ${formatBytes(backup.dbSize)}`);
+      yield* Effect.log(`  Database size: ${formatFileSize(backup.dbSize)}`);
     }
     if (backup.filesSize != null) {
       yield* Effect.log(
-        `  Storage files size: ${formatBytes(backup.filesSize)}`,
+        `  Storage files size: ${formatFileSize(backup.filesSize)}`,
       );
     }
     if (backup.expiresAt) {
